fix: 🐛 Cadastro de usuário corrigido, mudança do nome da userController para authController e adição de verificação no login
